Cache log handler state per instance
We can't easily use Redis in order to determine whether or not we've added a
log handler per pod, because of the fact that pods may restart, and if we've
marked a handler as being created we won't add it when the pod comes back up.

In our current logger [1] we cache the actual logger object, because that's what
we call. Watchtower adds a handler onto the existing log object, so we just need
to ensure the handler exists so we don't add it superfluously.

We'll restrict the log streams to those configured in the environment.

[1] https://github.com/RedHatInsights/platform-logger/blob/master/platform_logger.rb