Conditionally apply log handlers for different targets
Based on `LOG_TO_SPLUNK` and `LOG_TO_CLOUDWATCH` environment variables, when a log
request comes in, we'll add the log stream to the cache as we do today, but then
conditionally add handlers for Splunk and/or CloudWatch based on the environment
config.

We don't use the stream for Splunk, so in theory we could probably get away with
a single logger, but this keeps things separate in the event we want to isolate
logs by index/stream name in Splunk as well in the future. Right now we'll only
be using this for the gateway.