Add CA with a configMap conditionally
This will allow a CA to be added via a `trusted-certs` configMap (optionally), which
will add the CA to `/etc/ssl/certs/${TRUSTED_CERT_FILENAME}` based on the configMap
key/value.

In the event the configMap doesn't exist, this will not require the volume, but will
create a superfluous empty dir in `/etc/ssl/certs/` - long term we may want to consider
an alternate conditional approach, but since mounts aren't optional this may
be required to get a custom CA in environments where it's required.