Use raw request data vs parsing the JSON data
For Splunk, we need to send the raw JSON string in order for it to be consumed
correctly, otherwise the object will end up being encoded incorrectly.
This change ensures that will happen.

For Watchtower (CloudWatch), existing functionality will be preserved because of
the fact that the library will check to see if the payload is a dictionary before
attemtping to convert the object back to a string [1] which in this case will now
be a no-op.

[1] https://github.com/kislyuk/watchtower/blob/develop/watchtower/__init__.py#L134-L135